refactored footer
