+++ lib/src/views/ui/Pit.dart
@@ -19,7 +19,7 @@ class Pit extends StatelessWidget {
           children: [
             const HeaderTitle(),
             HeaderNav(currentPage: Pages.pit, changePage: changePage,),
-            const Footer(),
+            const Footer(pageTitle: "Pit"),
           ],
         )
     );
